Se agrego el endpoint GetProductsFilteredClients y se quito el endpoint GetProducts

+++ Dsw2025Tpi.Application/Dtos/DashboardModel.cs
@@ -10,8 +10,8 @@ namespace Dsw2025Tpi.Application.Dtos
     {
         public record Summary
         (
-            long productCount,
-            long orderCount
+            int productCount,
+            int orderCount
         );
     }
 }
